# lib/api-spec/openapi.yml
openapi: 3.1.0
info:
  title: TestCaseHub API
  version: "1.0.0"
  description: UAT Test Case Management API
servers:
  - url: http://localhost:3000/api
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Defect:
      type: object
      properties:
        id: { type: integer }
        test_run_id: { type: integer }
        test_case_id: { type: integer }
        execution_id: { type: integer }
        tester_notes: { type: string, nullable: true }
        status: { type: string }
        retest_reason: { type: string, nullable: true }
        accepted_by_business_note: { type: string, nullable: true }
        rejection_log: { type: string, nullable: true }
        severity: { type: string, nullable: true }
        priority: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DefectRetest:
      type: object
      properties:
        id: { type: integer }
        defect_id: { type: integer }
        test_run_id: { type: integer }
        assigned_tester_id: { type: integer, nullable: true }
        retest_result: { type: string, nullable: true }
        retest_notes: { type: string, nullable: true }
        retested_by_user_id: { type: integer, nullable: true }
        retested_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
...
paths:
...
  /test-runs/{testRunId}/defects:
    get:
      summary: List defects for test run
      tags: [Defects]
      parameters:
        - in: path
          name: testRunId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Defect"
  /defects/{defectId}:
    get:
      summary: Get defect by id
      tags: [Defects]
      parameters:
        - in: path
          name: defectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Defect"
  /defects/{defectId}/flag-bug:
    patch:
      summary: Flag defect as bug
      tags: [Defects]
      parameters:
        - in: path
          name: defectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /defects/{defectId}/flag-retest:
    patch:
      summary: Flag defect for retest
      tags: [Defects]
      parameters:
        - in: path
          name: defectId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          description: OK
  /defects/{defectId}/flag-accepted-by-business:
    patch:
      summary: Flag defect accepted by business
      tags: [Defects]
      parameters:
        - in: path
          name: defectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /defects/{defectId}/business-accept:
    patch:
      summary: Business owner accepts defect
      tags: [Defects]
      parameters:
        - in: path
          name: defectId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string }
      responses:
        "200":
          description: OK
  /defects/{defectId}/business-reject:
    patch:
      summary: Business owner rejects defect
      tags: [Defects]
      parameters:
        - in: path
          name: defectId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          description: OK
  /defects/{defectId}/classify:
    patch:
      summary: Classify defect
      tags: [Defects]
      parameters:
        - in: path
          name: defectId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                severity: { type: string }
                priority: { type: string }
      responses:
        "200":
          description: OK
  /defects/{defectId}/notes:
    post:
      summary: Add note to defect
      tags: [Defects]
      parameters:
        - in: path
          name: defectId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string }
      responses:
        "201":
          description: Created
  /defects/{defectId}/retests:
    get:
      summary: List retests for defect
      tags: [Defects]
      parameters:
        - in: path
          name: defectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DefectRetest"
...
  /test-runs/{testRunId}/discussions:
    post:
      summary: Create team discussion
      tags: [TeamDiscussions]
      parameters:
        - in: path
          name: testRunId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                meetingType: { type: string, enum: [defect_review, post_mortem] }
                participantIds: { type: array, items: { type: integer } }
      responses:
        "201":
          description: Created
  /discussions/{discussionId}:
    get:
      summary: Get discussion by id
      tags: [TeamDiscussions]
      parameters:
        - in: path
          name: discussionId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /discussions/{discussionId}/participants:
    post:
      summary: Add participant to discussion
      tags: [TeamDiscussions]
      parameters:
        - in: path
          name: discussionId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId: { type: integer }
                canAddNotes: { type: boolean }
      responses:
        "201":
          description: Created
  /discussions/{discussionId}/participants/{userId}:
    delete:
      summary: Remove participant from discussion
      tags: [TeamDiscussions]
      parameters:
        - in: path
          name: discussionId
          required: true
          schema:
            type: integer
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
  /discussions/{discussionId}/end:
    patch:
      summary: End discussion
      tags: [TeamDiscussions]
      parameters:
        - in: path
          name: discussionId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /discussions/{discussionId}/defects/{defectId}:
    get:
      summary: Get defects in discussion
      tags: [TeamDiscussions]
      parameters:
        - in: path
          name: discussionId
          required: true
          schema:
            type: integer
        - in: path
          name: defectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
...
security:
  - BearerAuth: []